Reconfigure the project files
