+++ grails-app/controllers/com/synacy/shoppingCenter/ShopController.groovy
@@ -1,4 +1,4 @@
-package shopping.center
+package com.synacy.shoppingCenter
 
 class ShopController {
