+++ grails-app/controllers/com/synacy/shoppingCenter/TagController.groovy
@@ -1,4 +1,4 @@
-package shopping.center
+package com.synacy.shoppingCenter
 
 class TagController {
