+++ grails-app/domain/com/synacy/shoppingCenter/Tag.groovy
@@ -1,4 +1,4 @@
-package com.synacy.shoppingcenter
+package com.synacy.shoppingCenter
 
 class Tag {
