+++ test/unit/com/synacy/shoppingCenter/ShopControllerSpec.groovy
@@ -1,4 +1,4 @@
-package shopping.center
+package com.synacy.shoppingCenter
 
 import grails.test.mixin.TestFor
 import spock.lang.Specification
@@ -6,8 +6,8 @@ import spock.lang.Specification
 /**
  * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
  */
-@TestFor(TagController)
-class TagControllerSpec extends Specification {
+@TestFor(ShopController)
+class ShopControllerSpec extends Specification {
 
     def setup() {
     }
